backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
networkx
python-dotenv
google-generativeai
python-dotenv==1.0.1
